move TestCase class TestDnaFreqCalcMethods from test_dna_utils module to new module just for DnaFreqCalc class method testing
