+++ base/tests/test_dna_utils.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
 """
-Test module for dna_utils classes within the bioinformatics-learning package
+Test module for the DnaUtil class within the bioinformatics-learning package
 
 file_utils unittest.TestCase classes:
-* Test TestDnaUtilsMethods
-
-Todo:
-    * Test additional classes added to solve course prompts
+* TestDnaUtilsMethods
 
 .. _Google Python Style Guide:
     http://google.github.io/styleguide/pyguide.html
@@ -16,8 +13,6 @@ import os
 import unittest
 
 from base.dna_utils import DnaUtil
-from base.dna_utils import DnaFreqCalc
-from base.file_utils import FileUtil
 from base.tests.fixtures import FIXTURE_PATH
 
 GOOD_DNA = 'ACGTACGT'
@@ -30,6 +25,7 @@ FIXTURE1_FNAME = 'SamplePatternCount.txt'
 FIXTURE2_FNAME = 'PatternCount.txt'
 FIXTURE1_FPATH = os.path.join(FIXTURE_PATH, FIXTURE1_FNAME)
 FIXTURE2_FPATH = os.path.join(FIXTURE_PATH, FIXTURE2_FNAME)
+
 
 class TestDnaUtilsMethods(unittest.TestCase):
     """
@@ -72,32 +68,4 @@ class TestDnaUtilsMethods(unittest.TestCase):
         dna_util_obj = DnaUtil(TEST_FWD_SEQ)
         expected = TEST_REV_COMP_SEQ
         actual = dna_util_obj.reverse_complement()
-        self.assertEqual(expected, actual)
-
-class TestDnaFreqCalcMethods(unittest.TestCase):
-    """
-    Test methods of DnaFreqCalc class from course examples
-    """
-    def test_pattern_count_sample(self):
-        """
-        Test the pattern_count method using the SamplePatternCount.txt inputs
-        """
-        file_util_obj = FileUtil(FIXTURE1_FPATH)
-        inputs = file_util_obj.inputs
-        expected = int(file_util_obj.outputs[0])
-        sequence_obj = DnaFreqCalc(inputs[0])
-        actual = sequence_obj.pattern_count(inputs[1])
-        self.assertEqual(expected, actual)
-
-    def test_pattern_count_extra_dataset(self):
-        """
-        Test the pattern_count method using the PatternCount.txt fixture
-
-        The course prompt called this the "Extra Dataset"
-        """
-        file_util_obj = FileUtil(FIXTURE2_FPATH)
-        inputs = file_util_obj.inputs
-        expected = int(file_util_obj.outputs[0])
-        sequence_obj = DnaFreqCalc(inputs[0])
-        actual = sequence_obj.pattern_count(inputs[1])
         self.assertEqual(expected, actual)
